Modified the recursive_assign interface and set_vocabs
