+++ xnmt/encoder.py
@@ -1,9 +1,10 @@
 import dynet as dy
 import model
-import decorators
 import embedder
 import serializer
 import model_globals
+
+from decorators import recursive
 
 # All types of encoder
 import residual
@@ -50,7 +51,7 @@ class LSTMEncoder(BuilderEncoder, Serializable):
     else:
       self.builder = dy.VanillaLSTMBuilder(layers, input_dim, hidden_dim, model)
 
-  @decorators.recursive
+  @recursive
   def set_train(self, val):
     self.builder.set_dropout(self.dropout if val else 0.0)
 
@@ -67,7 +68,7 @@ class ResidualLSTMEncoder(BuilderEncoder, Serializable):
     else:
       self.builder = residual.ResidualRNNBuilder(layers, input_dim, hidden_dim, model, dy.VanillaLSTMBuilder, residual_to_output)
 
-  @decorators.recursive
+  @recursive
   def set_train(self, val):
     self.builder.set_dropout(self.dropout if val else 0.0)
 
@@ -80,7 +81,7 @@ class PyramidalLSTMEncoder(BuilderEncoder, Serializable):
     self.dropout = dropout
     self.builder = pyramidal.PyramidalRNNBuilder(layers, input_dim, hidden_dim, model_globals.dynet_param_collection.param_col, dy.VanillaLSTMBuilder, downsampling_method, reduce_factor)
 
-  @decorators.recursive
+  @recursive
   def set_train(self, val):
     self.builder.set_dropout(self.dropout if val else 0.0)
 
@@ -95,7 +96,7 @@ class ConvBiRNNBuilder(BuilderEncoder, Serializable):
     self.builder = conv_encoder.ConvBiRNNBuilder(layers, input_dim, hidden_dim, model, dy.VanillaLSTMBuilder,
                                             chn_dim, num_filters, filter_size_time, filter_size_freq, stride)
 
-  @decorators.recursive
+  @recursive
   def set_train(self, val):
     self.builder.set_dropout(self.dropout if val else 0.0)
 
@@ -113,7 +114,7 @@ class ModularEncoder(Encoder, Serializable):
       sent = module.transduce(sent)
     return sent
 
-  @decorators.recursive
+  @recursive
   def set_train(self, val):
     for module in self.modules:
       module.set_train(val)
